requirements.txt
# Core Framework Dependencies
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3

# Authentication & Security
flask-login==0.6.3
flask-bcrypt==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & ORM
alembic==1.12.1
flask-sqlalchemy==3.1.1

# API & HTTP
requests==2.31.0
httpx==0.25.2
python-multipart==0.0.6

# AI Integration
google-generativeai==0.3.2
groq==0.4.1

# Excel Export
openpyxl==3.1.2
pandas==2.1.4

# Environment & Configuration
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# CORS
flask-cors==4.0.0